rearrange crowdin code into a project api class and a utils class
